requirements.txt
altgraph==0.17.2
awscli-local==0.20
boto3==1.24.5
botocore==1.27.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
gspread==5.4.0
httplib2==0.20.4
idna==3.3
jmespath==1.0.0
localstack-client==1.35
macholib==1.16
modulegraph==0.19.2
oauthlib==3.2.0
protobuf==4.21.2
py2app==0.28.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.6.0
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
